had dinos come back to life when I called buildCards(), issue fixed by splicing the original array of dinos
